Black_Box_main/requirements.txt
numpy
pandas
tqdm
pillow
torchvision
torch==1.9.0+cu100
argparse
os